feat: succinct 1.79

+++ library/std/src/sys/pal/zkvm/abi.rs
@@ -1,11 +1,4 @@
-//! ABI definitions for symbols exported by risc0-zkvm-platform.
-
-// Included here so we don't have to depend on risc0-zkvm-platform.
-//
-// FIXME: Should we move this to the "libc" crate?  It seems like other
-// architectures put a lot of this kind of stuff there.  But there's
-// currently no risc0 fork of the libc crate, so we'd either have to
-// fork it or upstream it.
+//! ABI definitions for symbols exported by succinct-zkvm-platform.
 
 #![allow(dead_code)]
 pub const DIGEST_WORDS: usize = 8;
@@ -19,7 +12,7 @@ pub mod fileno {
 }
 
 extern "C" {
-    // Wrappers around syscalls provided by risc0-zkvm-platform:
+    // Wrappers around syscalls provided by succinct-zkvm-platform:
     pub fn sys_halt();
     pub fn sys_output(output_id: u32, output_value: u32);
     pub fn sys_sha_compress(
@@ -34,7 +27,7 @@ extern "C" {
         buf: *const u8,
         count: u32,
     );
-    pub fn sys_rand(recv_buf: *mut u32, words: usize);
+    pub fn sys_rand(recv_buf: *mut u8, words: usize);
     pub fn sys_panic(msg_ptr: *const u8, len: usize) -> !;
     pub fn sys_log(msg_ptr: *const u8, len: usize);
     pub fn sys_cycle_count() -> usize;
